+++ usr/src/sys/kern/vfs_vnops.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)vfs_vnops.c	7.44 (Berkeley) %G%
+ *	@(#)vfs_vnops.c	7.45 (Berkeley) %G%
  */
 
 #include "param.h"
@@ -342,7 +342,7 @@ vn_ioctl(fp, com, data, p)
 		if (com == FIONREAD) {
 			if (error = VOP_GETATTR(vp, &vattr, p->p_ucred, p))
 				return (error);
-			*(off_t *)data = vattr.va_size - fp->f_offset;
+			*(int *)data = vattr.va_size - fp->f_offset;
 			return (0);
 		}
 		if (com == FIONBIO || com == FIOASYNC)	/* XXX */
